Move load_and_preprocess_image to the catgrad-llm lib
