+++ catgrad-llm/examples/siglip/main.rs
@@ -6,9 +6,8 @@ use catgrad::prelude::*;
 use catgrad::stdlib::nn::*;
 use catgrad::typecheck::TypeExpr;
 use catgrad_llm::helpers::*;
-use catgrad_llm::utils::{get_model_files, load_model_weights};
+use catgrad_llm::utils::{get_model_files, load_and_preprocess_image, load_model_weights};
 use clap::Parser;
-use image::imageops::FilterType;
 use std::path::PathBuf;
 use tokenizers::Tokenizer;
 
@@ -468,39 +467,6 @@ impl Module<2, 2> for SiglipModel {
 
         [logits_per_text, logits_per_image]
     }
-}
-
-// Loads the image and returns flattened data + shape
-fn load_and_preprocess_image(
-    image_path: &PathBuf,
-    image_size: usize,
-    patch_size: usize,
-) -> (Vec<f32>, Vec<usize>) {
-    let num_channels = 3;
-
-    let img = image::open(image_path).unwrap();
-    let resized_img =
-        img.resize_to_fill(image_size as u32, image_size as u32, FilterType::Triangle);
-    let rgb_img = resized_img.to_rgb8();
-    let img = rgb_img.into_raw();
-
-    let pixels: Vec<f32> = img.iter().map(|&x| x as f32 * (2. / 255.0) - 1.).collect();
-    // For image sizes 384x384 we need to truncate to 378x378 so it's a multiple of patch size.
-    let aligned_image_size = (image_size / patch_size) * patch_size;
-    let mut patches = vec![0.0; num_channels * aligned_image_size * aligned_image_size];
-    for row in 0..aligned_image_size {
-        for col in 0..aligned_image_size {
-            for chan in 0..num_channels {
-                patches[chan * aligned_image_size * aligned_image_size
-                    + row * aligned_image_size
-                    + col] = pixels[(row * image_size + col) * num_channels + chan];
-            }
-        }
-    }
-    (
-        patches,
-        vec![1, num_channels, aligned_image_size, aligned_image_size],
-    )
 }
 
 pub fn main() -> Result<(), Box<dyn std::error::Error>> {
