+++ catgrad-llm/src/utils.rs
@@ -489,3 +489,39 @@ pub fn load_model<B: interpreter::Backend>(
         total_params,
     ))
 }
+
+// Loads the image and returns flattened data + shape
+pub fn load_and_preprocess_image(
+    image_path: &PathBuf,
+    image_size: usize,
+    patch_size: usize,
+) -> (Vec<f32>, Vec<usize>) {
+    let num_channels = 3;
+
+    let img = image::open(image_path).unwrap();
+    let resized_img = img.resize_to_fill(
+        image_size as u32,
+        image_size as u32,
+        image::imageops::FilterType::Triangle,
+    );
+    let rgb_img = resized_img.to_rgb8();
+    let img = rgb_img.into_raw();
+
+    let pixels: Vec<f32> = img.iter().map(|&x| x as f32 * (2. / 255.0) - 1.).collect();
+    // For image sizes 384x384 we need to truncate to 378x378 so it's a multiple of patch size.
+    let aligned_image_size = (image_size / patch_size) * patch_size;
+    let mut patches = vec![0.0; num_channels * aligned_image_size * aligned_image_size];
+    for row in 0..aligned_image_size {
+        for col in 0..aligned_image_size {
+            for chan in 0..num_channels {
+                patches[chan * aligned_image_size * aligned_image_size
+                    + row * aligned_image_size
+                    + col] = pixels[(row * image_size + col) * num_channels + chan];
+            }
+        }
+    }
+    (
+        patches,
+        vec![1, num_channels, aligned_image_size, aligned_image_size],
+    )
+}
